requirements.txt
asgiref==3.4.1
Django==3.2.7
Pillow==8.4.0
pytz==2021.1
sqlparse==0.4.2